move options to codec.compression package
