+++ codec-http/src/test/java/io/netty/handler/codec/http/BrotliContentCompressorTest.java
@@ -16,7 +16,6 @@
 package io.netty.handler.codec.http;
 
 import io.netty.channel.embedded.EmbeddedChannel;
-import io.netty.handler.codec.http.compression.StandardCompressionOptions;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
